Uploaded First Test
I have written a python program that detects several driving factors and responds to them accordingly. 
Turning and braking have been the focus on this project as I don’t yet have a computer. 
- The turning aspect works by calculating the middle of the lane (not the frame) and calculating the difference in that point compared the middle point of the driver(frame in this tests case) 
- The braking aspect works similarly by calculating the difference of 100 and in approximated ft; the distance of another vehicle. 
They both work in tandem and almost everything is hard-coded besides of the cascade for vehicle detection.